Create og tooltip component
